requirements.txt
fastapi
uvicorn
openai
python-dotenv
psycopg2
semantic-kernel==1.28.1